refactor(platform): extract IsWSL to internal/platform; decouple config from audio
internal/config imported internal/audio solely for IsWSL (per review
finding #68). Move IsWSL + detectWSLFromData + readProcVersion +
truncateString to a new internal/platform stdlib-only leaf package.

internal/config no longer imports internal/audio. The audio package
keeps no platform-detection responsibility -- the WSL probe is a
generic OS check, not an audio concern.

Five config call sites + four config-test sites + one CLI call site
updated. internal/audio/factory.go and DetectOptimalBackend now
import internal/platform.

Resolves review finding #68.
