+++ internal/audio/platform.go
@@ -2,46 +2,10 @@ package audio
 
 import (
 	"log/slog"
-	"os"
 	"os/exec"
-	"strings"
+
+	"claudio.click/internal/platform"
 )
-
-// IsWSL checks if the current environment is Windows Subsystem for Linux
-func IsWSL() bool {
-	return detectWSLFromData(readProcVersion(), os.Getenv("WSL_DISTRO_NAME"))
-}
-
-// detectWSLFromData checks for WSL indicators in the provided data (for testing)
-func detectWSLFromData(procVersion, wslEnv string) bool {
-	slog.Debug("checking WSL detection", "proc_version_snippet", truncateString(procVersion, 50), "wsl_env", wslEnv)
-
-	// Check WSL_DISTRO_NAME environment variable (WSL sets this)
-	if wslEnv != "" {
-		slog.Debug("WSL detected via environment variable", "distro", wslEnv)
-		return true
-	}
-
-	// Check /proc/version for Microsoft or WSL indicators
-	procLower := strings.ToLower(procVersion)
-	if strings.Contains(procLower, "microsoft") || strings.Contains(procLower, "wsl") {
-		slog.Debug("WSL detected via /proc/version", "indicators", "microsoft or wsl found")
-		return true
-	}
-
-	slog.Debug("no WSL indicators found")
-	return false
-}
-
-// readProcVersion reads /proc/version file content
-func readProcVersion() string {
-	content, err := os.ReadFile("/proc/version")
-	if err != nil {
-		slog.Debug("failed to read /proc/version", "error", err)
-		return ""
-	}
-	return string(content)
-}
 
 // CommandExists checks if a command is available in the system's PATH using exec.LookPath
 func CommandExists(command string) bool {
@@ -57,7 +21,7 @@ func CommandExists(command string) bool {
 
 // DetectOptimalBackend determines the best audio backend for the current system
 func DetectOptimalBackend() string {
-	return detectOptimalBackendWithChecker(IsWSL(), CommandExists)
+	return detectOptimalBackendWithChecker(platform.IsWSL(), CommandExists)
 }
 
 // detectOptimalBackendWithChecker allows dependency injection for testing
@@ -102,12 +66,4 @@ func getPreferredSystemCommandWithChecker(commandChecker func(string) bool) stri
 
 	slog.Debug("no preferred system audio commands found")
 	return ""
-}
-
-// truncateString truncates a string to maxLen characters for logging
-func truncateString(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	return s[:maxLen] + "..."
 }
